requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
cmsplugin-contact==1.0.0
django-classy-tags==0.3.4.1
django-cms==2.2
django-mptt==0.5.2
django-sekizai==0.5
django-tinymce==1.5.1b2
html5lib==0.90
psycopg2==2.4.4
wsgiref==0.1.2
